+++ Changes.md
@@ -14,7 +14,8 @@
 ### Code Improvements
 
 1. Got rid of the depricated scriptableUnescapeHTML and hence incompatibility with Easy YouTube Video Downloader
-
+2. Used iccb for IITK.CSE.CS213.BYTubeD everywhere.
+3. Centralized stripHTML
 
 ## 1.1.1